# Node48 add: make it SIMD, fully rolled for now