This introduces SSE4.1 intrinsics over the previous SSE2 ones, add a note and a
compilation flag.

Baseline performance:

$ ./micro_benchmark_node48 --benchmark_filter=add
2020-10-29T05:35:22+01:00
Running ./micro_benchmark_node48
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.22, 0.09, 0.02
-------------------------------------------------------------------------------------
Benchmark                           Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------
node48_sequential_add/2          6.55 us         6.54 us       107032 items_per_second=9.47406M/s size=12.1895k
node48_sequential_add/8          23.3 us         23.3 us        30011 items_per_second=10.6661M/s size=48.6953k
node48_sequential_add/64          255 us          255 us         2743 items_per_second=8.14183M/s size=402.013k
node48_sequential_add/512        2341 us         2341 us          299 items_per_second=7.17849M/s size=3.16333M
node48_sequential_add/4096      19537 us        19535 us           36 items_per_second=6.88069M/s size=25.3126M
node48_random_add/2              6.05 us         6.04 us       115634 items_per_second=10.266M/s size=12.1895k
node48_random_add/8              21.3 us         21.3 us        32915 items_per_second=11.6685M/s size=48.6953k
node48_random_add/64              206 us          206 us         3388 items_per_second=10.0689M/s size=402.013k
node48_random_add/512            2318 us         2318 us          302 items_per_second=7.24918M/s size=3.16333M
node48_random_add/4096          22993 us        22991 us           30 items_per_second=5.84648M/s size=25.3126M

With the patch:

$ ./micro_benchmark_node48 --benchmark_filter=add
2020-10-29T05:58:18+01:00
Running ./micro_benchmark_node48
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.94, 0.32, 0.11
-------------------------------------------------------------------------------------
Benchmark                           Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------
node48_sequential_add/2          5.87 us         5.87 us       118794 items_per_second=10.569M/s size=12.1895k
node48_sequential_add/8          20.6 us         20.6 us        33935 items_per_second=12.03M/s size=48.6953k
node48_sequential_add/64          239 us          239 us         2939 items_per_second=8.69739M/s size=402.013k
node48_sequential_add/512        2205 us         2205 us          317 items_per_second=7.61996M/s size=3.16333M
node48_sequential_add/4096      18455 us        18453 us           38 items_per_second=7.28409M/s size=25.3126M
node48_random_add/2              5.18 us         5.18 us       135261 items_per_second=11.9739M/s size=12.1895k
node48_random_add/8              18.2 us         18.2 us        38529 items_per_second=13.6471M/s size=48.6953k
node48_random_add/64              191 us          190 us         3662 items_per_second=10.9103M/s size=402.013k
node48_random_add/512            2213 us         2213 us          316 items_per_second=7.59291M/s size=3.16333M
node48_random_add/4096          22337 us        22336 us           31 items_per_second=6.01779M/s size=25.3126M